backend/requirements.txt
async-timeout==5.0.1
asyncpg==0.31.0
et_xmlfile==2.0.0
greenlet==3.2.4
openpyxl==3.1.5
python-dotenv==1.2.1
SQLAlchemy==2.0.46
typing_extensions==4.15.0
